requirements/base.txt
django==3.1.2
djangorestframework
markdown
django-filter
phonenumbers==8.12.13
pillow==8.0
django-cors-headers